Remove extraneous whitespace; make attributes consistent
This puts the attributes into a consistent ordering. The attributes
were in a kind of order, but not all of them were ordered in the same
way, so about twenty files had to have the order of attributes
changed.

The other change in this commit is that the inconsistent use of
whitespace in the XML files was removed. This was in about 72
files. For example some files had a space after the word "Commons" in
the licence text, or some files had two spaces between attributes, or
some files had blank lines or wrong indentation.

The purpose of this commit is to make it easier to operate on the
files using software which reads XML and then writes it out again.