Fix problems with ito
The variant forms in kanji_mismatch containing an ito component 糸
seem to have all been created by a script which had an off-by-one
error in the number of strokes. Compounding the error, these were all
grouped incorrectly. This commit shifts the groupings to put six
rather than five strokes into the groups of the itos, and shifts the
subsequent child paths of the entire kanji as well.