Elements made consistent
This changes kvg:element value of the top level SVG group containing
the paths to be the same as the Unicode value taken from the file
name.

kanji/04f34-HzFst.svg contained both incorrect element and incorrect
path type information so this was edited to put approximately correct
path types and groups as well.