Remove some instances where "nelson" was misapplied
The five files in this commit all have a general radical in the
correct place plus a "nelson" radical in somewhere very unlikely, but
which uses the same original element as the general radical. Although
I don't have a copy of Nelson to check against, I'm fairly certain
these are errors, since Nelson was at least consistent.