Change bad path type in 𠕁 subelements
This change consists of searching for 𠕁 elements with IDS, then
searching the KanjiVG file for a group with element 冊, then looking
at this group's paths to see if they correspond to the type, and
swapping them around if so. The single-element groups on these paths
were in a mess already, so I have just removed them completely.

From the groupings, it seems as if this was done to add another stroke
order variant for this 冊 piece, but in fact all of the found elements
in KanjiVG had the same stroke order.