Remove the duplicate radicals
This uses a script to remove all of the duplicate radicals. I spot
checked the geometry, and it seems that the first group is always the
right one to keep, because of the stroke order, so I just removed all
the other radical values on groups except the first one. I did not
actually go to the trouble of sorting the geometries in order since
there didn't seem to be any exceptions to the "first is upper left"
rule.