Add "nyoc" grouping, meaning "nyo counterpart"
Most of this is automatically edited. About 300 cases were simple
addition of the nyoc group to whichever of a pair of groups lacked the
position element, then about 50 or so cases were re-grouping of the
counterpart elements into a single group and adding the nyoc position,
and then there were a few cases such as 051a0.svg of grouping errors
in the existing files, where the nyo counterpart had falsely been made
a child of the nyo group which had to be hand edited.

There were no other types of errors as far as I could see, and this
now covers every case of "nyo" position in the kanji directory.