Doubled cases of contiguous groups
This does some more cases where the file has two joined groups.

The script which does this work is not very sophisticated, so it was
necessary to run it twice for cases there were two instances of groups
which should be joined.

I also ran it a third time, but there were no cases of three sets of
groups like this.