Change 月 to 円 and remove subgroup
This automatically changes a number of cases where the bottom part of
an 青 element was written using an 円 kanji but marked as 月 in the
group with a subgroup of 円 to using a single group with 月 as the
original value and 円 as the element value, without a subgroup.

None of these groups was a radical or otherwise seemed to require
special treatment.