Change left stroke of 月 to consistent values
Where the left stroke was labelled ㇑/㇒, it's been relabelled as ㇑,
CJK Stroke S. This label seems to have been intended to indicate that
the stroke was vertical but could be bent in some cases, but its used
was inconsistent, so I decided to remove it.

Where the left stroke was labelled ㇒, two actions were taken
depending on the properties of the stroke.

Where the stroke appeared to bend left at the bottom, the stroke was
relabelled ㇓, CJK Stroke SP, which is the correct stroke type for
this according to the documentation available.

Where the stroke appeared to be straight at the bottom and vertical,
the stroke was relabelled ㇑, CJK Stroke S. (Some of the strokes had a
small kink at the top.)

All of the files in the second case had "Kaisho" in the file name,
except for one exception, 0745c-VtLst.svg.