Consistently make the upper dot of 白 U+31D4
This changes every case of 白 in KanjiVG to have the dash, "CJK STROKE
D", as its upper stroke.

I think most of the cases here were ones I had changed in error
earlier not knowing that the dash should be used for these cases, but
there also seem to be some which were there initially.