Fix positions/groups on 04f5e.svg
Whoever made the KanjiVG groups originally often insisted on plugging
in ill-matching groups like 仁 in characters like these. I don't want
to remove these willy-nilly but if these things must be added, at
least they should be added so they don't disrupt the actual structure
of the character. This splits the 仁 bit into two parts and makes it
no longer the "master group" of this character.