Change 09083 kanji groupings
This makes the groupings into road+other rather than hole+other.

Also, the odd 遂 grouping was not actually applied to the "road"
element from strokes 15 to 19, so I've added it over the road element,
as well as lowering its position in the hierarcy as low as possible
over the upper element, since this is not actually a good
representation of the character visually.

See https://github.com/KanjiVG/kanjivg/issues/374 for discussion.