Add groups for 敫𦥯㳟𦔮𡗗耑电翏夌
This commit consists of multiple additions of groups on empty elements
from the list 敫𦥯㳟𦔮𡗗耑电翏夌. Two characters, 09675.svg, with an
incorrect group element, and 07ac5.svg, where the white element was
covering more strokes than it should, were edited in a text editor to
correct the mistakes. The rest of the edits were automated.