Change ㇑ to ㇚ where appropriate in Kaisho variants
Many of the Kaisho variants consist of nothing other than a down-line
in a tree shape with a hook at the end, but this difference is not
represented in the stroke's kvg:type. This uses a graphical search
algorithm to find instances where a line is labelled ㇑, but appears
to actually be ㇚ graphically, and automatically changes the files.

There are about 150 changes. I've done spot checks on about ten of the
files to make sure that the correct things are being done.