A large number of automatic edits; see previous commit
The script which checks the errors in the strokes is very slow & also
requires clustering of results. It also has to be run over and over to
find these errors, since once too many wrong things get into the
cluster of results it can't discriminate any more. This represents
about three runs.