Automated fix of a large number of swapped stroke type errors
This edit was mostly done automatically, with some spot checks and one
hand edit. It consists of looking for pairs of horizontal and vertical
strokes with the types the wrong way around in the "*HzFst*" and
similarly named files. There are about eighty examples. I have
spot-checked the changes on about ten cases.