Fix a common problem with 虎 swapped strokes
There are a lot of files with Hz/VtFstRiLe names where the third and
fourth strokes were swapped. Many of these also have Nelson radicals
based on the incorrect stroke. This swaps them around.

This is an automated edit and not every change was checked by human. I
checked about five characters in total afterwards.