Change the type of the sixth stroke of 足
These mostly occur in -Kaisho files where the sixth stroke of a 足
element has the type ㇒, presumably because there was a plan to draw
it that way, but the element doesn't actually have that type since it
wasn't drawn in the sloping form. This automatically changes all
instances to ㇑. A few spot checks were performed.