Repair mistaken horizontals/verticals in 曲
This edit is an automated repair of a large number of -Kaisho variants
with a stroke type within 曲 which is horizontal where it should be
vertical, or vice-versa. Some of the cases seemed to be swapped, and
some of them were just with a single line in the wrong direction.