Replace ㇑ with ㇔ using geometry
This adds to previous edits where a vertical was replaced with a dash
for "Kaisho" files. In all cases the line didn't appear to be a
vertical from its geometry, and formed part of a "hat" structure. The
judgement on what was vertical was made using an average over all
verticals in the KanjiVG data set, and a cut off of being over 3.3
standard deviations away from the mean of the vertical lines was used
as the criterion. Spot checks were performed on about five cases of
the total of about 500 changes. The previous similar edits were done
not with the statistics, but with numbers I had typed in.