Swap of CJK Stroke T/P
This is an automated edit with about five of the changes
spot-checked. This deals with a large number of cases where CJK Stroke
T and CJK Stroke P have been used for a stroke going in the wrong
direction. The assumption of this edit is that the graphic is correct
and the type is wrong where the two disagree. Where a horizontal
stroke is also a possibility, the change is skipped over, since many
mistaken cases seem to be CJK Stroke H rather than either of the two
slanting lines.