Automated change of ㇐ to ㇀/㇐
This is an automated change of altering ㇐ to ㇀/㇐ in cases where the
stroke slants upward, typically in cases where the stroke is
horizontal in the original element but slants upward in KanjiVG when
used as a subelement. About ten spot checks have been carried out, and
no errors were noticed.